refactor: build a sequence of formula rather than a monolithic one
Two reasons:
- Building a single monolithic formula may have the unfortunate effect
of adding individual formulas to the bdd incrementally, rather than building a
bdd for a given formula and taking the conjunction with the existing
bdd. It shouldn't happen, according to the `compile_to_bdd`
function. But now, I'm sure!
- It will help change the order in which the formulas are inserted in
the bdd. In order to try and reduce the number of choices in the bdd
and, hopefully, drastically reduce the memory usage.